workspaces/naver-blog/scripts/requirements.txt
Pillow>=10.0.0
argparse
